Operator buttons don't un-shorten display number
In effect, I converted the display to a number in the function, instead
of converting it when operator buttons are pressed, and had the function
round everything it returns.

I could've rounded the display number when the operator buttons are
pressed, but it's best practice to not have to do that outside the
function everywhere it's called. Also, I didn't realize that I could've
done that until after I did everything ;P